Moving factory package to dto package.
